Bug fix for aws hopefully: Changed the data that is passed around to UTC on front and back
The frontend should be sending a better formatted string to the back. The back end should now be storing the Kind field as UTC on creation. Then on getEvents, it needs to be specified the kind is UTC for the browser to display correctly. There might be a way to build this into EF core database directly, to where it says everything is UTC always