P2(filtering) basic work done
It is working on a very basic level to where P2 is done, but there are some user interactions that I don't like and want to fix if I have time later. Such as the filtering resets on page reload or page changes which is expected to happen, but I want to see if there's a way to change that and save state. So if a user selects or unselects a filter, it's saved when they come back to the page later
Also make it so all calendars are selected by default for displaying on first launch