Can't delete group event from My Schedule page, can only do on group page
I decided to do this since users could accidentally delete group events on their individual calendar. To me, it made sense to prevent this to avoid some confusion with users and to avoid accidentally deletions of group events from a user's personal schedule.
If we decide we don't this to be how it works, we can revert this change